scraper_framework/requirements.txt
# Web Scraping Framework Requirements
# Core scraping dependencies
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.0
scrapy==2.11.0
lxml==4.9.3
html5lib==1.1

# Async and performance
aiohttp==3.9.1
asyncio==3.4.3
concurrent-futures==3.1.1

# Data processing
pandas==2.1.4
numpy==1.24.3
python-dateutil==2.8.2

# Database and storage
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
redis==5.0.1

# Task scheduling
celery==5.3.4
celery[redis]==5.3.4

# Anti-detection and proxies
fake-useragent==1.4.0
requests-html==0.10.0
undetected-chromedriver==3.5.4
selenium-stealth==1.0.6
rotating-proxies==0.6.2

# Data validation and parsing
pydantic==2.5.1
jsonschema==4.20.0
python-dotenv==1.0.0

# Logging and monitoring
structlog==23.2.0
sentry-sdk==1.38.0

# Utilities
tqdm==4.66.1
click==8.1.7
python-slugify==8.0.1
fuzzywuzzy==0.18.0
python-levenshtein==0.23.0

# Image processing (for captcha handling)
pillow==10.1.0
opencv-python==4.8.1.78

# Machine learning for data cleaning
scikit-learn==1.3.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1